volume/shim: drop it
The volume shim driver was introduced in commit 61113e3. It's supposed
to help plugin authors focus on the volume driver API, rather than the
plugin API. However, this means the plugins SDK is pulling in
github.com/docker/docker.

Since moby/moby@04f5276, the Engine's volume package has been moved
into the daemon package, and is considered an implementation detail that
no one should import.

Looking at the use of `NewHandlerFromVolumeDriver` on grep.app, and
Sourcegraph, it seems no one is using it. So drop it unceremoniously.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>